Code refactoring; improved docstrings
