--- src/extraction/features/analytics/__init__.py
@@ -1,1 +1,0 @@
-"""Analytics module for tracking matching parameter usage."""
